# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          }
        },
        "summary": "GET /notifications — 본인 알림 목록 (미읽음 우선/최신순)",
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        },
        "summary": "PATCH /notifications/:id/read — 본인 알림 읽음 처리",
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/admin/banners": {
      "post": {
        "operationId": "AdminBannerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBannerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BannerResponse"
                }
              }
            }
          }
        },
        "summary": "POST /admin/banners — 배너 생성",
        "tags": [
          "AdminBanner"
        ]
      },
      "get": {
        "operationId": "AdminBannerController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BannerResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "GET /admin/banners — 전체 목록 (관리자)",
        "tags": [
          "AdminBanner"
        ]
      }
    },
    "/admin/banners/{id}": {
      "patch": {
        "operationId": "AdminBannerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBannerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BannerResponse"
                }
              }
            }
          }
        },
        "summary": "PATCH /admin/banners/:id — 배너 부분 수정",
        "tags": [
          "AdminBanner"
        ]
      },
      "delete": {
        "operationId": "AdminBannerController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "DELETE /admin/banners/:id — 배너 삭제",
        "tags": [
          "AdminBanner"
        ]
      }
    },
...
    "/admin/sellers/{id}/approve": {
      "post": {
        "operationId": "AdminController_approveSeller",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerProfileResponse"
                }
              }
            }
          }
        },
        "summary": "POST /admin/sellers/:id/approve — 판매자 승인 (seller 도메인 재사용) + 감사 로그 기록",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            }
          }
        },
        "summary": "GET /admin/users — 사용자 목록 (cursor 페이지네이션)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/audit-logs": {
      "get": {
        "operationId": "AdminController_listAuditLogs",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminAuditLogResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "GET /admin/audit-logs — 관리자 조치 감사 로그 (최신순, 013)",
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "info": {
    "title": "DOA Market API",
    "description": "DOA Market 백엔드 HTTP 계약 (프론트 코드젠 SSOT).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token (재발급)"
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "description": "cross-schema plain String — users.users.id (P-001)"
          },
          "type": {
            "enum": [
              "ORDER_PLACED",
              "ORDER_SHIPPED",
              "SETTLEMENT_CREATED",
              "REVIEW_RECEIVED"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "title",
          "body",
          "isRead",
          "createdAt"
        ]
      },
      "NotificationListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "size": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size"
        ]
      },
...
      "CreateBannerDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "linkUrl": {
            "type": "string",
            "description": "클릭 이동 대상 URL (없으면 비링크 배너)"
          },
          "position": {
            "type": "string",
            "description": "노출 위치 (미지정 시 기본 MAIN_TOP)",
            "enum": [
              "MAIN_TOP",
              "MAIN_MIDDLE",
              "MAIN_BOTTOM",
              "SIDEBAR"
            ]
          },
          "sortOrder": {
            "type": "number",
            "description": "동일 위치 내 정렬 순서 (오름차순)"
          },
          "isActive": {
            "type": "boolean",
            "description": "활성 여부 (미지정 시 기본 true)"
          },
          "startsAt": {
            "type": "string",
            "description": "노출 시작 (ISO 8601, null=즉시)"
          },
          "endsAt": {
            "type": "string",
            "description": "노출 종료 (ISO 8601, null=무제한)"
          }
        },
        "required": [
          "title",
          "imageUrl"
        ]
      },
      "BannerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "linkUrl": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "enum": [
              "MAIN_TOP",
              "MAIN_MIDDLE",
              "MAIN_BOTTOM",
              "SIDEBAR"
            ],
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "startsAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "imageUrl",
          "position",
          "sortOrder",
          "isActive",
          "createdAt"
        ]
      },
      "UpdateBannerDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "linkUrl": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "string",
            "enum": [
              "MAIN_TOP",
              "MAIN_MIDDLE",
              "MAIN_BOTTOM",
              "SIDEBAR"
            ]
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "startsAt": {
            "type": "string",
            "nullable": true
          },
          "endsAt": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SellerProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "description": "cross-schema plain String — users.users.id (P-001)"
          },
          "businessName": {
            "type": "string"
          },
          "businessNumber": {
            "type": "string"
          },
          "representativeName": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string",
            "nullable": true
          },
          "businessAddress": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ],
            "type": "string"
          },
          "rejectReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "businessName",
          "businessNumber",
          "representativeName",
          "status",
          "createdAt"
        ]
      },
      "AdminUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt"
        ]
      },
      "AdminUserListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
      "AdminAuditLogResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "adminId": {
            "type": "string",
            "description": "조치 수행 관리자 userId (P-001)"
          },
          "action": {
            "type": "string",
            "description": "조치 종류 (예: SELLER_APPROVE)"
          },
          "targetType": {
            "type": "string",
            "description": "대상 엔티티 종류 (예: SELLER, BANNER, USER)"
          },
          "targetId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "adminId",
          "action",
          "targetType",
          "targetId",
          "createdAt"
        ]
      }
    }
  }
}